맞춤서비스 추천 리스트 뜸 / 새로고침 문제

+++ app/src/main/java/com/example/subsmanager2/entity/RecommendEntity.kt
@@ -8,7 +8,7 @@ data class RecommendEntity(    //
     var planName:String     = "",
     var subsName:String     = "",
     var fee:String          = "",
-    var boardRating:Double  = 0.0,
+    var boardRating:Float  = 0.0F,
     var imgUrl:String       = ""
 ) {
     @Exclude
